Enable kafka_consumer to restart for SSL file changes
