+++ common/kafka/kafka_consumer.h
@@ -15,44 +15,14 @@
 #pragma once
 
 #include <atomic>
-#include <map>
-#include <memory>
 #include <string>
 #include <unordered_set>
 
 #include "common/MultiFilePoller.h"
-#include "gflags/gflags.h"
-#include "glog/logging.h"
-#include "librdkafka/rdkafkacpp.h"
+#include "common/kafka/kafka_consumer_holder.h"
+
 
 namespace kafka {
-
-class RdKafkaConsumerHolder {
-public:
-  RdKafkaConsumerHolder() {};
-
-  // no copy nor move
-  RdKafkaConsumerHolder(const RdKafkaConsumerHolder&) = delete;
-
-  RdKafkaConsumerHolder(RdKafkaConsumerHolder&&) = delete;
-
-  virtual ~RdKafkaConsumerHolder() {}
-
-  virtual std::shared_ptr<RdKafka::KafkaConsumer> getInstance() = 0;
-  virtual void resetInstance() = 0;
-  virtual void close() = 0;
-};
-
-class RdKafkaConsumerHolderFactory {
-public:
-  static RdKafkaConsumerHolder* createInstance(const std::unordered_set <uint32_t> &partition_ids,
-                                                const std::string &broker_list,
-                                                const std::unordered_set <std::string> &topic_names,
-                                                const std::string &group_id,
-                                                const std::string &kafka_consumer_type);
-
-  static RdKafkaConsumerHolder* createInstance(std::shared_ptr <RdKafka::KafkaConsumer> consumer);
-};
 
 class KafkaConsumer {
 public:
@@ -126,6 +96,7 @@ private:
   std::shared_ptr<common::MultiFilePoller::CallbackId> cbIdPtr_;
   std::atomic<bool> is_healthy_;
   std::atomic<bool> reset_;
+  TopicPartitionToValueMap<ConsumedOffset> consumedOffsets_;
 };
 
 }  // namespace kafka
